pyproject.toml
[project]
name = "swin2sr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
]
